requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
PyMuPDF==1.24.4
pydantic==2.7.4
google-generativeai==0.5.4
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
pandas==2.2.2
streamlit==1.36.0
orjson==3.10.3
